requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-cloud-storage==2.16.0
google-cloud-firestore==2.16.0
google-genai>=0.2.0 
pdfplumber==0.11.0
pytesseract==0.3.10
python-multipart==0.0.9
PyMuPDF==1.26.3
firebase-admin==6.5.0
httpx==0.27.0
python-dotenv==1.0.1
numpy
fpdf2==2.7.6
websockets
pdf2image==1.17.0
Pillow==10.3.0